Move operation utils into a single module
