--- strawberry/utils/get_first_operation.py
@@ -1,13 +1,0 @@
-from typing import Optional
-
-from graphql.language import DocumentNode, OperationDefinitionNode
-
-
-def get_first_operation(
-    graphql_document: DocumentNode,
-) -> Optional[OperationDefinitionNode]:
-    for definition in graphql_document.definitions:
-        if isinstance(definition, OperationDefinitionNode):
-            return definition
-
-    return None
